+++ pkgs/html/CHANGELOG.md
@@ -1,8 +1,12 @@
-## 0.15.7-wip
+## 0.15.7
 
+- `writeTextNodeAsHtml`: Escape text inside `<script>` and `<style>`-like tags in foreign namespaces.
+- Fix the tokenizer failing to close CDATA blocks when encountering extra brackets
 - Require Dart `3.6`
+- Limit ["Noah's Ark" clause](https://html.spec.whatwg.org/multipage/parsing.html#list-of-active-formatting-elements) to avoid degenerate O(N^2) when parsing deeply nested formatting elements with unique attributes.
 - Added new `textContent` method to `Node` class that returns the text content of
   the node.
+- Fix XSS vulnerability in `htmlToCodeMarkup()` by escaping DOCTYPE, element, and attribute names.
 
 ## 0.15.6